👌 IMP: Refactor time management
